requirements.txt
python-telegram-bot==22.1
Pillow==11.2.1
requests==2.31.0
python-dotenv==1.1.0
